support long

+++ qqtextwar/src/main/java/cn/qqtextwar/blocks/Block.java
@@ -6,11 +6,11 @@ public class Block {
 
     private Vector vector;
 
-    private int id;
+    private long id;
 
     private boolean cross;
 
-    public Block(Vector vector, int id, boolean cross) {
+    public Block(Vector vector,long id, boolean cross) {
         this.vector = vector;
         this.id = id;
         this.cross = cross;
@@ -20,7 +20,7 @@ public class Block {
         return vector;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
